+++ app/db/Cache.scala
@@ -12,6 +12,8 @@ import models.{ SensorModel, DatapointModel }
 trait Cache {
   def calculateBinsByYear(sensor_id: Int, since: Option[String], until: Option[String], parameter: String)
 
+  def calculateBinsBySeason(sensor_id: Int, since: Option[String], until: Option[String], parameter: String)
+
   def calculateBinsByMonth(sensor_id: Int, since: Option[String], until: Option[String], parameter: String)
 
   def calculateBinsByDay(sensor_id: Int, since: Option[String], until: Option[String], parameter: String)
@@ -20,6 +22,8 @@ trait Cache {
 
   def insertIntoBinYear(sensor: SensorModel, dp: DatapointModel, yyyy: Int, parameter: String, value: Double)
 
+  def insertIntoBinSeason(sensor: SensorModel, dp: DatapointModel, yyyy: Int, season: String, parameter: String, value: Double)
+
   def insertIntoBinMonth(sensor: SensorModel, dp: DatapointModel, yyyy: Int, mm: Int, parameter: String, value: Double)
 
   def insertIntoBinDay(sensor: SensorModel, dp: DatapointModel, yyyy: Int, mm: Int, dd: Int, parameter: String, value: Double)
@@ -27,6 +31,8 @@ trait Cache {
   def insertIntoBinHour(sensor: SensorModel, dp: DatapointModel, yyyy: Int, mm: Int, dd: Int, hh: Int, parameter: String, value: Double)
 
   def getCachedBinStatsByYear(sensor: SensorModel, since: Option[String], until: Option[String], parameter: String, total: Boolean): List[(Int, Int, Double, Double, Timestamp, Timestamp)]
+
+  def getCachedBinStatsBySeason(sensor: SensorModel, season: Option[String], since: Option[String], until: Option[String], parameter: String, total: Boolean): List[(Int, String, Int, Double, Double, Timestamp, Timestamp)]
 
   def getCachedBinStatsByMonth(sensor: SensorModel, since: Option[String], until: Option[String], parameter: String, total: Boolean): List[(Int, Int, Int, Double, Double, Timestamp, Timestamp)]